+++ app/src/serializers/fieldSerializer.js
@@ -9,8 +9,8 @@ class FieldSerializer {
     }
 
     static serialize(data, tableName) {
-        if (data && data.length > 0) {
-            const { mappings } = data[0][tableName];
+        if (data && data[tableName]) {
+            const { mappings } = data[tableName];
             if (mappings.type) {
                 return {
                     tableName,